Lay the groundwork for refactoring the Flask backend
